producer/requirements.txt
flask>=2.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.0
pywebpush>=2.0.0
py-vapid>=1.9.0
python-dotenv>=1.0.0
requests>=2.28.0
